Set round details attempt
First attempt at setting rounds from the onGameStart callback. Doesn't seem to work as currently done (for this I think you would need to set attributes in the addRound() function. Such as game.addRound(name: x, type: y).  For now will try to move to onRoundStart callback as this is likely better and easier to read.