Update starting positions
Current approach is to have players clumped near eachother with (very approximately, particularly in the high universilazibility condition) the same number of steps to the nearest water source, and with a bit of a walk to get there (nobody too close). Need to think more carefully about this, and maybe try to exactly control this. Also, another interesting experiment idea, would be to have players play consistently in the same map but get randomly generated in different positions. It is a very literal way way to look at manipulating 'original position" in veil of ignorance reasoning approach. See how this changes behaviour (maybe the lining up norm develops quickly here even in repeated games on the same map as players realise that they have no idea what their original position will be).